Changed register to hold a byte, fixed setting the carry flag in sub

+++ TitanEmulator/Register.cs
@@ -5,11 +5,11 @@ using System.Text;
 
 namespace TitanEmulator {
     public class Register {
-        public int value;
+        public byte value;
         public int index;
 
         public Register(int v, int i) {
-            value = v;
+            value = (byte) v;
             index = i;
         }
     }
