+++ TitanEmulator/instructions/AdcInstruction.cs
@@ -9,7 +9,7 @@ namespace TitanEmulator.instructions {
         }
 
         public override void execute(MachineState ms, byte[] parameters) {
-            ms.registers[parameters[0]].value = ms.registers[parameters[0]].value + ms.registers[parameters[1]].value + ms.flags["C"];
+            ms.registers[parameters[0]].value =(byte) ( ms.registers[parameters[0]].value + ms.registers[parameters[1]].value + ms.flags["C"]);
         }
     }
 }
