+++ TitanEmulator/instructions/NotInstruction.cs
@@ -10,7 +10,7 @@ namespace TitanEmulator.instructions {
 
         public override void execute(MachineState ms, byte[] parameters) {
             int res = ~ms.registers[parameters[0]].value & 0xFF;
-            ms.registers[parameters[0]].value = res;
+            ms.registers[parameters[0]].value = (byte) (res);
             ms.flags["S"] = res >> 4 > 0x8 ? 1 : 0;
             ms.flags["Z"] = res == 0 ? 1 : 0;
         }
